+++ src/views/Vacancies/components/index.ts
@@ -1,3 +1,3 @@
 export { default as FilterSidebar } from "./FilterSidebar";
-export { default as FilterTagsHeader } from "./FilterTagsHeader";
+export { default as ActiveFilterTags } from "./ActiveFilterTags";
 export { default as VacancyCard } from "./VacancyCard";
